+++ gcp-test/tests/gcp_bigquery_table/test-get-query.sql
@@ -1,0 +1,18 @@
+select 
+  table_id,
+  dataset_id, 
+  id, 
+  kind, 
+  labels, 
+  range_partitioning, 
+  time_partitioning, 
+  type, 
+  self_link,  
+  title, 
+  location, 
+  project
+from 
+  gcp.gcp_bigquery_table
+where 
+  table_id = '{{ resourceName }}'
+  and dataset_id = '{{ resourceName }}';