+++ gcp-test/tests/gcp_compute_resource_policy/test-invalid-name-query.sql
@@ -1,0 +1,3 @@
+select name, id, description
+from gcp.gcp_compute_resource_policy
+where name = '';