+++ gcp-test/tests/gcp_compute_resource_policy/test-list-query.sql
@@ -1,0 +1,3 @@
+select name, description
+from gcp.gcp_compute_resource_policy
+where title = '{{ resourceName }}';