+++ gcp-test/tests/gcp_dns_policy/test-turbot-query.sql
@@ -1,0 +1,7 @@
+select
+  title,
+  akas
+from
+  gcp_dns_policy
+where
+  name = '{{ resourceName }}';