+++ gcp-test/tests/gcp_dns_record_set/test-invalid-name-query.sql
@@ -1,0 +1,3 @@
+select name, type
+from gcp.gcp_dns_record_set
+where name = '';