+++ gcp-test/tests/gcp_logging_bucket/test-turbot-query.sql
@@ -1,0 +1,3 @@
+select title, akas, location, project
+from gcp.gcp_logging_bucket
+where name = '{{ resourceName }}' and location = '{{ output.region_id.value }}';