+++ gcp-test/tests/gcp_project/test-list-query.sql
@@ -1,0 +1,8 @@
+select
+  name,
+  project_id,
+  project_number
+from
+  gcp_project
+where
+  project_id = '{{ output.project_id.value }}';