+++ gcp-test/tests/gcp_project_organization_policy/test-turbot-query.sql
@@ -1,0 +1,3 @@
+select title, akas
+from gcp.gcp_project_organization_policy
+where title = '{{ output.resource_title.value }}';