+++ gcp/table_gcp_sql_database_instance_metric_connections_daily.go
@@ -1,0 +1,48 @@
+package gcp
+
+import (
+	"context"
+
+	"github.com/turbot/steampipe-plugin-sdk/grpc/proto"
+	"github.com/turbot/steampipe-plugin-sdk/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/plugin/transform"
+	sqladmin "google.golang.org/api/sqladmin/v1beta4"
+)
+
+//// TABLE DEFINITION
+
+func tableGcpSQLDatabaseInstanceMetricConnectionsDaily(_ context.Context) *plugin.Table {
+	return &plugin.Table{
+		Name:        "gcp_sql_database_instance_metric_connections_daily",
+		Description: "GCP SQL Database Instance Metrics - Connections (Daily)",
+		List: &plugin.ListConfig{
+			ParentHydrate: listSQLDatabaseInstances,
+			Hydrate:       listSQLDatabaseInstanceMetricConnectionsDaily,
+		},
+		Columns: monitoringMetricColumns([]*plugin.Column{
+			{
+				Name:        "instance_id",
+				Description: "The ID of the instance.",
+				Type:        proto.ColumnType_STRING,
+				Transform:   transform.FromField("DimensionValue"),
+			},
+		}),
+	}
+}
+
+//// LIST FUNCTION
+
+func listSQLDatabaseInstanceMetricConnectionsDaily(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
+	instanceInfo := h.Item.(*sqladmin.DatabaseInstance)
+
+	// Get project details
+	projectData, err := activeProject(ctx, d)
+	if err != nil {
+		return nil, err
+	}
+	project := projectData.Project
+	location := instanceInfo.Region
+	dimensionValue := "\"" + project + ":" + instanceInfo.Name + "\""
+
+	return listMonitorMetricStatistics(ctx, d, "DAILY", "\"cloudsql.googleapis.com/database/network/connections\"", "resource.label.database_id = ", dimensionValue, instanceInfo.Name, location)
+}